ectool-cbi: Check the input size properly
Currently, the input value size is assigned to a uint8_t. This does not
allow us to reject a large value because it won't exceed 255.

This CL changes the variable to size_t.

BUG=b:311452202
TEST=Build

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5038868
Tested-by: Daisuke Nojiri <dnojiri@chromium.org>
Auto-Submit: Daisuke Nojiri <dnojiri@chromium.org>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>

+++ util/ectool.cc
@@ -8774,7 +8774,7 @@ static int cmd_cbi(int argc, char *argv[])
 			(struct ec_params_set_cbi *)ec_outbuf;
 		void *val_ptr;
 		uint64_t val = 0;
-		uint8_t size;
+		size_t size;
 		uint8_t bad_size = 0;
 		if (argc < 5) {
 			fprintf(stderr, "Invalid number of params\n");
@@ -8809,7 +8809,7 @@ static int cmd_cbi(int argc, char *argv[])
 					bad_size = 1;
 			}
 			if (bad_size == 1) {
-				fprintf(stderr, "Bad size: %d\n", size);
+				fprintf(stderr, "Bad size: %zu\n", size);
 				return -1;
 			}
 
@@ -8817,7 +8817,7 @@ static int cmd_cbi(int argc, char *argv[])
 		}
 
 		if (size > ec_max_outsize - sizeof(*p)) {
-			fprintf(stderr, "Size exceeds parameter buffer: %d\n",
+			fprintf(stderr, "Size exceeds parameter buffer: %zu\n",
 				size);
 			return -1;
 		}
